add types for Canvas.tsx
