+++ src/components/Header.tsx
@@ -6,7 +6,7 @@ import cx from "classnames"
 
 import Massick1x1 from "../svg/massick-1x1.svg"
 
-import Transition from "./transition"
+import Transition from "./Transition"
 import Language from "./Language"
 
 const navLinks = [
